requirements.txt
gradio==4.26.0
onnx==1.12.0
onnxruntime-training==1.15.1
tensorflow==2.10.1
tf2onnx==1.14.0
protobuf==3.19.0
numpy~=1.26.4
pillow~=10.3.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113